+++ src/systems/si/units/mod.rs
@@ -6,13 +6,13 @@ mod scaled;
 
 pub use base::{Ampere, Candela, Kelvin, Kilogram, Meter, Mole, Radian, Second};
 pub use derived::{
-    Coulomb, CubicMeter, CubicMeterPerSecond, Farad, Gray, GrayPerSecond, Hertz, Joule,
-    JoulePerCubicMeter, JoulePerKelvin, JoulePerKilogramKelvin, JoulePerMole, JoulePerMoleKelvin,
-    JoulePerSquareMeter, KilogramPerCubicMeter, KilogramPerCubicMeterSecond, MeterPerSecond,
-    Newton, NewtonPerMeter, Ohm, Pascal, PascalSecond, PerCubicMeter, PerKelvin, PerMeter,
-    PerSecond, PerSquareKelvin, Rayl, Siemens, SquareMeter, SquareMeterPerKilogram,
-    SquareMeterPerSecond, Volt, Watt, WattPerCubicMeter, WattPerKilogram, WattPerMeterKelvin,
-    WattPerSquareMeter,
+    Coulomb, CoulombPerSquareMeter, CubicMeter, CubicMeterPerSecond, Farad, Gray, GrayPerSecond,
+    Hertz, Joule, JoulePerCubicMeter, JoulePerKelvin, JoulePerKilogramKelvin, JoulePerMole,
+    JoulePerMoleKelvin, JoulePerSquareMeter, KilogramPerCubicMeter, KilogramPerCubicMeterSecond,
+    KilogramPerSecond, MeterPerSecond, MeterPerVolt, Newton, NewtonPerMeter, Ohm, Pascal,
+    PascalPerVolt, PascalSecond, PerCubicMeter, PerKelvin, PerMeter, PerSecond, PerSquareKelvin,
+    Rayl, Siemens, SquareMeter, SquareMeterPerKilogram, SquareMeterPerSecond, Volt, VoltPerPascal,
+    Watt, WattPerCubicMeter, WattPerKilogram, WattPerMeterKelvin, WattPerSquareMeter,
 };
 pub use scaled::{
     Centimeter, CubicMillimeter, ElectronVolt, Gram, GramPerCubicCentimeter, JoulePerMilliliter,